==> 2.1-Uniprot-Data-Parser.ipynb <==
### Bulk Parsing and Saving

import gzip
import json
import pandas as pd
import xml.etree.ElementTree as ET
import os
from parser.single_xml_parser import UniprotEntryParser
from parser.splited_xml_parser import SplitedEntryParser

### Preparing sample function

def parse_file(file_path): 
    tree = ET.parse(file_path)
    root = tree.getroot()

    SEP = SplitedEntryParser(root)

    entry_dict = {'accession': SEP.get_accession(),\
                   'name': SEP.get_name(),\
                    'gene': SEP.get_gene(),\
                    'organism': SEP.get_organism(),\
                    'sequence': SEP.get_sequence(),\
                    'uniprotId': SEP.get_uniprotId(),\
                    'ptm': SEP.get_ptm(),\
                    'references': SEP.get_references()}
        
    return entry_dict

parse_file('data/xmldataset/entry_546400.xml')

### Counting files in the folder and getting the file paths

file_dir = 'xmldataset'
os.listdir(file_dir)[0]

nfiles = 10
file_paths = [os.path.join(file_dir, file) for file in os.listdir(file_dir)[0:nfiles] if file.endswith('.xml')]
file_paths[0]

---------------------------------------------------------------------------

### Parallelize the parsing of xml files
Run file from terminal: ```python3 parallel.py```

Parsed data will be dumped as ```all_uniprot_data.json```

---------------------------------------------------------------------------

### Mapping the Reference into PTM Evidences

with open("output/all_uniprot_data.json", "r") as f:
    uniprot_data = json.load(f) 

for item in uniprot_data:
    key2reference = {}
    for reference in item['references']:
        key2reference[reference['key']] = reference

with open('kgdata/key2reference.json', 'w') as f:
    json.dump(key2reference, f)
