==> 3.2-Uniprot-KG-Data-Builder.ipynb <==
### Uniprot KG Data Builder

import gzip
import pandas as pd
import json as json
import os


#### Read Json file (one file element at a time)

with open("data/output/all_uniprot_data.json", "r") as f:
    uniprot_data = json.load(f) 

- How amny data points are there? What are the keys?

print("Data Points: ",len(uniprot_data),"|",  "Keys:" , uniprot_data[0].keys())

- Lets print single data that includes PTM info- 

import pprint
pp = pprint.PrettyPrinter(indent=4)
pp.pprint(uniprot_data[24])

- What are other PTM including data?

for i,item in enumerate(uniprot_data):
    if item['ptm'] != []:
        print(item['ptm'])
        
    if i >50:
        break   


### Prepare subset of data with PTM record

First we select the PTM associated data and flatten the dictionary to create three columns seperately: Position, description, evicence

PTM_DATA = []
for item in uniprot_data:
    if item['ptm'] != []:
        data = {}
        data['accession'] = item['accession']
        data['gene'] = item['gene']
        data['name'] = item['name']
        data['organism'] = item['organism']
        data['sequence'] = item['sequence']
        data['uniprotId'] = item['uniprotId']
        

        idata = data.copy()
        
        for ptm in item['ptm']:
            try:
                idata['position'] = ptm['Position']
            except:
                idata['position'] = None
            try:
                idata['description'] = ptm['Description']
            except:
                idata['description'] = None
            try:
                idata['evidence'] = ptm['Evidence'].split(' ')
            except:
                idata['evidence'] = None

            PTM_DATA.append(idata.copy())

# with open('output/PTM_DATA.json', 'w') as f:
#     json.dump(PTM_DATA,f)

#with open('output/PTM_DATA.json', 'r') as f:
#     PTM_DATA = json.load(f)

- How many PTM data are found? What is the PTM data %?

len(PTM_DATA), (len(PTM_DATA)/len(uniprot_data))*100

df_ptm_flat = pd.DataFrame(PTM_DATA)

df_ptm_flat.head(2)

#### Create PTM Identifier from description:
WE want to create identifier from description column to match identifier from PTM vocab file


df_ptm_flat['Identifier'] = df_ptm_flat['description']\
        .apply(lambda x: x.split(';')[0].lower() if ";" in x else x.lower())

with open('data/ptm/idf2accession.json', 'r') as f:
    idf2accession = json.load(f)

df_ptm_flat['ptm_accession'] = df_ptm_flat['Identifier']\
        .apply(lambda x: idf2accession[x] if x in idf2accession else None)

df_ptm_flat.head(2)

df_ptm_flat.columns

#260688
df_ptm_flat_dropna = df_ptm_flat.dropna(subset=['ptm_accession'])
len(df_ptm_flat_dropna), (len(df_ptm_flat_dropna)/len(df_ptm_flat))*100

df_ptm_flat_dropna['mod_res_id'] = df_ptm_flat_dropna['uniprotId'] \
        + "_" + df_ptm_flat_dropna['ptm_accession'] + "_" + df_ptm_flat_dropna['position'].astype(str)

df_ptm_flat_dropna.head(5)

# save the dataframe as csv file
#df_ptm_flat_dropna.to_csv('kgdata/ptm_flat_dropna.csv', index=False)

# save the dataframe as json file
df_ptm_flat_dropna.to_json('data/kgdata/ptm_flat_dropna.json', orient='records')

df_ptm_flat_dropna.columns

# Rough

# explode the dictionaries into separate rows
#df_test = pd.read_csv('output/ptm_data.csv')
#df_test = df_test.explode('ptm')
#df_test.head(5)

#convert the dictionaries into separate columns
#df_test = pd.read_csv('output/ptm_data.csv')
#df_test[['Position', 'Description', 'Evidence']] = pd.json_normalize(df_test['ptm'])
