==> 4.1-Uniprot-KG-Builder.ipynb <==
## Uniprot KG Builder

import pandas as pd
import numpy as np
import os
import time
import neo4j as neo4j
from neo4j import GraphDatabase

### Make a connection with Graph DataBase


driver=GraphDatabase.driver(uri="bolt://127.0.0.1:7687",auth=("neo4j","12345678"))
session=driver.session()

driver

### Creat constraint

def create_constraints(driver):
        '''
        Create constraints for nodes in the graph
        args:
            driver: neo4j driver
        returns:
            None
        '''
        query = ["CREATE CONSTRAINT UniqueOrganismNameConstraint FOR (O:Organism) REQUIRE O.name IS UNIQUE",\
                "CREATE CONSTRAINT UniqueProteinIdConstraint FOR (P:Protein) REQUIRE P.UID IS UNIQUE",\
                "CREATE CONSTRAINT UniquePathwayIdConstraint FOR (PW:Pathway) REQUIRE PW.PWID IS UNIQUE",\
                "CREATE CONSTRAINT UniqueCompoundIdConstraint FOR (C:Compound) REQUIRE C.CID IS UNIQUE",\
                "CREATE CONSTRAINT UniqueModifiedResidueIdConstraint FOR (MR:Modified_residue) REQUIRE MR.MRID IS UNIQUE",\
                "CREATE CONSTRAINT UniquePTMTypeIdConstraint FOR (PTM:PTM_Type) REQUIRE PTM.PTMID IS UNIQUE",\
                "CREATE CONSTRAINT UniqueReferenceIdConstraint FOR (R:Reference) REQUIRE R.PMID IS UNIQUE"]
        
        
        with driver.session() as session:
            for constraint in query:
                session.run(constraint)

'''UNCOMMENT AND RUN THIS CELL ONLY ONCE'''
#create_constraints(driver)

### Populate Node

#### 1. Protein Node

- Columns in PTM flat file:
``` 
       Index(['accession', 'gene', 'name', 'organism', 'sequence', 'uniprotId',
       'position', 'description', 'evidence', 'Identifier', 'ptm_accession',
       'mod_res_id'],
      dtype='object')
```

# Create protein node
def create_protein_node(data):
        '''
        Create Protein node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (P:Protein{UID:value.uniprotId})\
            ON CREATE SET P.sequence=value.sequence,\
            P.name=value.name, P.gene=value.gene"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

     

t1 = time.time()
data = "ptm_flat_dropna.json"
create_protein_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Pathway Nodes

def create_pathway_node(data):
        '''
        Create Pathway node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (PW:Pathway{PWID:value.Pathway_identifier})\
            ON CREATE SET PW.name=value.Pathway_name"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

     

t1 = time.time()
data = "pathway_node.json"
create_pathway_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Compound node

# Create Pathway node
def create_compound_node(data):
        '''
        Create Pathway node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (C:Compound{CID:value.Identifier})\
            ON CREATE SET C.name=value.MoleculeName,\
            C.type=value.MoleculeType"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

     

t1 = time.time()
data = "pathways_molecules.json"
create_compound_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create Modified Resude Node

# Create modified residue node
def create_modified_residue_node(data):
        '''
        Create Protein node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None

        Index(['accession', 'gene', 'name', 'organism', 'sequence', 'uniprotId',
       'position', 'description', 'evidence', 'Identifier', 'ptm_accession',
       'mod_res_id'],
        dtype='object')
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (MR:Modified_residue{MRID:value.mod_res_id})\
            ON CREATE SET MR.position=value.position,\
            MR.parent_protein=value.uniprotId,\
            MR.description=value.description,\
            MR.ptm_accession=value.ptm_accession,\
            MR.identifier=value.Identifier,\
            MR.evidence=value.evidence"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

     

t1 = time.time()
data = "ptm_flat_dropna.json"
create_modified_residue_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

# Create modified residue node from cystine data with columns: 
# Organism,Uniprot_AC,position,standard_seq,PTM_Catalogy,PTM_Type,
# ref,cell_line,Identification_Strategy,Identification_Technique,LID

def create_cystine_modified_residue_node(data):
        '''
        Create cystine modified residue node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None

        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (MR:Modified_residue{MRID:value.LID})\
            ON CREATE SET MR.position=value.position,\
            MR.Uniprot_AC=value.Uniprot_AC"
            
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

     

t1 = time.time()
data = "aac_kg.json"
create_cystine_modified_residue_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

#### 2. Create PTM Node

- Columns in PTM vocab file:
``` 
       Index(['Identifier', 'Accession', 'Feature_key', 'Target',
       'Position_amino_acid', 'Position_polypeptide', 'Correction_formula',
       'Monoisotopic_mass_difference', 'Average_mass_difference',
       'Cellular_location', 'Taxonomic_range', 'Reference', 'Keyword'],
      dtype='object')
```

# create ptm node
def create_ptmtype_node(data):
        '''
        Create PTM_Type node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (PTM:PTM_Type{PTMID:value.Accession})\
            ON CREATE SET PTM.identifier=value.Identifier,\
            PTM.feature_key=value.Feature_Key,\
            PTM.target=value.Target,\
            PTM.position_amino_acid=value.Position_amino_acid,\
            PTM.position_polypeptide=value.Position_polypeptide,\
            PTM.correction_formula=value.Correction_formula,\
            PTM.monoisotopic_mass_difference=value.Monoisotopic_mass_difference,\
            PTM.average_mass_difference=value.Average_mass_difference,\
            PTM.keyword=value.Keyword,\
            PTM.cellular_location=value.Cellular_location,\
            PTM.taxonomic_range=value.Taxonomic_range,\
            PTM.reference=value.Reference"
            


            #print(query)
            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)
    

t1 = time.time()
data = "ptm-vocab.json"
create_ptmtype_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

# create ptm node with column values: 
# Organism, PTM_Catalogy,PTM_Type,
# ref,cell_line,Identification_Strategy,Identification_Technique


def create_cysteine_PTM_Type_node(data):
        '''
        Create PTM_Type node in the graph
        args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (PTM:PTM_Type{PTMID:value.PTM_Type})\
            ON CREATE SET PTM.PTM_Catalogy=value.PTM_Catalogy,\
            PTM.cell_line=value.cell_line,\
            PTM.Identification_Strategy=value.Identification_Strategy,\
            PTM.Identification_Technique=value.Identification_Technique,\
            PTM.identifier=value.PTM_Type"

            #print(query)
            tx.run(query,data=data)
            
        with driver.session() as session:   
            session.execute_write(tx_function,data)
    

t1 = time.time()
data = "aac_kg.json"
create_cysteine_PTM_Type_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

#### 3. Organism Node

# Create organism node
def create_organism_node(data):
        '''  
        Create Organism node in the graph
         args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (O:Organism{name:value.organism})"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)


t1 = time.time()
data = "ptm_flat_dropna.json"
create_organism_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create reference node *****

```
Index(['gene', 'name', 'organism', 'uniprotId', 'position', 'description',
       'evidence', 'pubmedId', 'journal', 'title', 'date', 'Identifier',
       'ptm_accession'],
      dtype='object')

```


# Create reference node
def create_reference_node(data):
        '''  
        Create reference node in the graph
         args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (R:Reference{PMID:value.pubmedId})\
            ON CREATE SET R.title=value.title,\
            R.journal=value.journal,\
            R.date=value.date"
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)



t1 = time.time()
data = "df_modres_ref.json"
create_reference_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

#,Organism,Uniprot_AC,position,standard_seq,PTM_Catalogy,PTM_Type,ref,cell_line,Identification_Strategy,\
#Identification_Technique,LID


# Create reference node
def create_cysteine_reference_node(data):
        '''  
        Create reference node in the graph
         args:
            data: dictionary containing the data for the node
        returns:
            None
        '''
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MERGE (R:Reference{PMID:value.ref})"
            
            
            #print(query)
            tx.run(query,data=data)
        
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "aac_kg.json"
create_cysteine_reference_node(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

## Edge creation

### PPI : Protein-Protein Interaction

def create_edge_ppi(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (P1:Protein{UID:value.IDA}),\
            (P2:Protein{UID:value.IDB})\
            MERGE (P1)-[r:HAS_PPI_INTERACTION]->(P2)"
            
            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)


t1 = time.time()
data = "ppi.json"
create_edge_ppi(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create relationship between Protein and Modified Residue

def create_edge_protein2modified_residue(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (P: Protein {UID:value.uniprotId})\
            MATCH (MR: Modified_residue {MRID:value.mod_res_id})\
            MERGE (P)-[:HAS_MODIFIED_RESIDUE]->(MR)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "ptm_flat_dropna.json"
create_edge_protein2modified_residue(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

def create_edge_protein2_cysteinemodified_residue(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (P: Protein {UID:value.Uniprot_AC})\
            MATCH (MR: Modified_residue {MRID:value.LID})\
            MERGE (P)-[:HAS_MODIFIED_RESIDUE]->(MR)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "aac_kg.json"
create_edge_protein2_cysteinemodified_residue(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create relationship between Modified residue and  PTM type

def create_edge_modified_residue2ptm_type(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (MR:Modified_residue {MRID:value.mod_res_id})\
            MATCH (PTM:PTM_Type {PTMID:value.ptm_accession})\
            MERGE (MR)-[:HAS_PTM_TYPE]->(PTM)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "ptm_flat_dropna.json"
create_edge_modified_residue2ptm_type(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

def create_edge_cysteine_modified_residue2ptm_type(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (MR:Modified_residue {MRID:value.LID})\
            MATCH (PTM:PTM_Type {PTMID:value.PTM_Type})\
            MERGE (MR)-[:HAS_PTM_TYPE]->(PTM)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "aac_kg.json"
create_edge_cysteine_modified_residue2ptm_type(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create edge protein to pathway

def create_edge_protein2pathway(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (P:Protein {UID:value.Submitted_entities_found})\
            MATCH (PW:Pathway {PWID:value.Pathway_identifier})\
            MERGE (P)-[:PARTICIPATES_IN]->(PW)"


            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "pathway_edge.json"
create_edge_protein2pathway(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Edge Compound to Pathway

def create_edge_compound2pathway(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (C:Compound {CID:value.Identifier})\
            MATCH (PW:Pathway {PWID:value.Pathway_identifier})\
            MERGE (C)-[:PARTICIPATES_IN]->(PW)"
            


            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)


t1 = time.time()
data = "pathways_molecules.json"
create_edge_compound2pathway(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create relationship between Protein and Organism

def create_edge_protein2organism(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (P: Protein {UID:value.uniprotId})\
            MATCH (O: Organism {name:value.organism})\
            MERGE (P)-[:BELONGS_TO]->(O)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "ptm_flat_dropna.json"
create_edge_protein2organism(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Create relationship between Modified Residue and Reference

def create_edge_modres2ref(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (MR:Modified_residue {MRID:value.mod_res_id})\
            MATCH (R:Reference {PMID:value.pubmedId})\
            MERGE (MR)-[:HAS_REFERENCE]->(R)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "df_modres_ref.json"
create_edge_modres2ref(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

def create_edge_cysteine_modres2ref(data):
        def tx_function(tx,data):
            query = "WITH '" + data  + "' as url \
            CALL apoc.load.json(url) YIELD value \
            MATCH (MR:Modified_residue {MRID:value.LID})\
            MATCH (R:Reference {PMID:value.ref})\
            MERGE (MR)-[:HAS_REFERENCE]->(R)"

            tx.run(query,data=data)
        with driver.session() as session:   
            session.execute_write(tx_function,data)

t1 = time.time()
data = "aac_kg.json"
create_edge_cysteine_modres2ref(data)
t2 = time.time()
print( "success! total time: ", t2-t1)

### Rough
